Added storage abstract class
