+++ README.md
@@ -109,7 +109,7 @@ KAFKA_PRODUCER='{"bootstrap_servers": "localhost:1234", "retries": 3}'
 
 - [x] Kafka Entrypoint
 - [x] Kafka Dependency
-- [ ] Advanced feature select commit strategies: 
+- [ ] Commit strategies: 
     - _ALMOST_ONCE_DELIVERY_
     - _AT_LEAST_ONCE_DELIVERY_ 
     - _EXACTLY_ONCE_DELIVERY_
