+++ nameko_kafka/commits/abc.py
@@ -1,0 +1,15 @@
+"""
+    Commits abstract base class
+"""
+
+from abc import ABC, abstractmethod
+
+
+class Commits(ABC):
+    """
+        Abstract base class for commits
+    """
+
+    @abstractmethod
+    def read(self, topic: str, partition: int) -> int:
+        pass
